source-redshift-batch: Add 'emit_sourced_schemas' feature flag
This involves a moderate amount of refactoring to the discovery
code, but nothing very surprising.

Currently sourced schemas are emitted only once at capture startup.
It would be nice to emit them more regularly during execution, but
that gets a bit complicated due to how the workers run, so this is
a good starting point.
